SCRUM-40 fixsonar quality gate problems
